requirements.txt
websockets
requests
python-dotenv
aiohttp
pytz
supabase
beautifulsoup4
better_profanity
aiohttp-session cryptography